+++ aerospike_suite_test.go
@@ -1,4 +1,4 @@
-// Copyright 2013-2020 Aerospike, Inc.
+// Copyright 2014-2021 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,7 +16,12 @@ package aerospike_test
 
 import (
 	"bytes"
+	"crypto/tls"
+	"crypto/x509"
+	"encoding/pem"
 	"flag"
+	"fmt"
+	"io/ioutil"
 	"log"
 	"math/rand"
 	"os"
@@ -29,6 +34,7 @@ import (
 
 	as "github.com/aerospike/aerospike-client-go"
 	asl "github.com/aerospike/aerospike-client-go/logger"
+	ast "github.com/aerospike/aerospike-client-go/types"
 )
 
 var (
@@ -40,6 +46,12 @@ var (
 	useReplicas = flag.Bool("use-replicas", false, "Aerospike will use replicas as well as master partitions.")
 	debug       = flag.Bool("debug", false, "Will set the logging level to DEBUG.")
 	namespace   = flag.String("n", "test", "Namespace")
+
+	certFile          = flag.String("cert_file", "", "Certificate file name.")
+	keyFile           = flag.String("key_file", "", "Key file name.")
+	keyFilePassphrase = flag.String("key_file_passphrase", "", "Key file's pass phrase.")
+	nodeTLSName       = flag.String("node_tls_name", "", "Node's TLS name.")
+	rootCA            = flag.String("root_ca", "", "Root certificate.")
 
 	clientPolicy *as.ClientPolicy
 	client       *as.Client
@@ -70,7 +82,13 @@ func initTestVars() {
 		clientPolicy.AuthMode = as.AuthModeExternal
 	}
 
-	client, err = as.NewClientWithPolicy(clientPolicy, *host, *port)
+	// setup TLS
+	clientPolicy.TlsConfig = initTLS()
+
+	dbHost := as.NewHost(*host, *port)
+	dbHost.TLSName = *nodeTLSName
+
+	client, err = as.NewClientWithPolicyAndHost(clientPolicy, dbHost)
 	if err != nil {
 		log.Fatal(err.Error())
 	}
@@ -121,6 +139,11 @@ func securityEnabled() bool {
 	return err == nil
 }
 
+func xdrEnabled() bool {
+	res := info(client, "get-config:context=xdr")
+	return len(res) > 0 && !strings.HasPrefix(res, "ERROR")
+}
+
 func nsInfo(ns string, feature string) string {
 	node := client.GetNodes()[0]
 	infoMap, err := node.RequestInfo(as.NewInfoPolicy(), "namespace/"+ns)
@@ -144,8 +167,110 @@ func info(client *as.Client, feature string) string {
 	node := client.GetNodes()[0]
 	infoMap, err := node.RequestInfo(as.NewInfoPolicy(), feature)
 	if err != nil {
-		log.Fatal("Failed to connect to aerospike: err:", err)
+		if ae, ok := err.(ast.AerospikeError); ok {
+			return ae.Error()
+		} else {
+			log.Fatal("Failed to connect to aerospike: err:", err)
+		}
 	}
 
 	return infoMap[feature]
+}
+
+func initTLS() *tls.Config {
+	if len(*rootCA) == 0 && len(*certFile) == 0 && len(*keyFile) == 0 {
+		return nil
+	}
+
+	// Try to load system CA certs, otherwise just make an empty pool
+	serverPool, err := x509.SystemCertPool()
+	if serverPool == nil || err != nil {
+		log.Printf("Adding system certificates to the cert pool failed: %s", err)
+		serverPool = x509.NewCertPool()
+	}
+
+	if len(*rootCA) > 0 {
+		// Try to load system CA certs and add them to the system cert pool
+		caCert, err := readFromFile(*rootCA)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		log.Printf("Adding CA certificate to the pool...")
+		serverPool.AppendCertsFromPEM(caCert)
+	}
+
+	var clientPool []tls.Certificate
+	if len(*certFile) > 0 || len(*keyFile) > 0 {
+
+		// Read cert file
+		certFileBytes, err := readFromFile(*certFile)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		// Read key file
+		keyFileBytes, err := readFromFile(*keyFile)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		// Decode PEM data
+		keyBlock, _ := pem.Decode(keyFileBytes)
+		certBlock, _ := pem.Decode(certFileBytes)
+
+		if keyBlock == nil || certBlock == nil {
+			log.Fatalf("Failed to decode PEM data for key or certificate")
+		}
+
+		// Check and Decrypt the the Key Block using passphrase
+		if x509.IsEncryptedPEMBlock(keyBlock) {
+			decryptedDERBytes, err := x509.DecryptPEMBlock(keyBlock, []byte(*keyFilePassphrase))
+			if err != nil {
+				log.Fatalf("Failed to decrypt PEM Block: `%s`", err)
+			}
+
+			keyBlock.Bytes = decryptedDERBytes
+			keyBlock.Headers = nil
+		}
+
+		// Encode PEM data
+		keyPEM := pem.EncodeToMemory(keyBlock)
+		certPEM := pem.EncodeToMemory(certBlock)
+
+		if keyPEM == nil || certPEM == nil {
+			log.Fatalf("Failed to encode PEM data for key or certificate")
+		}
+
+		cert, err := tls.X509KeyPair(certPEM, keyPEM)
+
+		if err != nil {
+			log.Fatalf("Failed to add client certificate and key to the pool: `%s`", err)
+		}
+
+		log.Printf("Adding client certificate and key to the pool...")
+		clientPool = append(clientPool, cert)
+	}
+
+	tlsConfig := &tls.Config{
+		Certificates:             clientPool,
+		RootCAs:                  serverPool,
+		InsecureSkipVerify:       false,
+		PreferServerCipherSuites: true,
+	}
+	tlsConfig.BuildNameToCertificate()
+
+	return tlsConfig
+}
+
+// Read content from file
+func readFromFile(filePath string) ([]byte, error) {
+	dataBytes, err := ioutil.ReadFile(filePath)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to read from file `%s`: `%v`", filePath, err)
+	}
+
+	data := bytes.TrimSuffix(dataBytes, []byte("\n"))
+
+	return data, nil
 }