+++ batch_command_exists.go
@@ -1,4 +1,4 @@
-// Copyright 2013-2020 Aerospike, Inc.
+// Copyright 2014-2021 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -45,7 +45,6 @@ func newBatchCommandExists(
 		keys:        keys,
 		existsArray: existsArray,
 	}
-	res.oneShot = false
 	return res
 }