+++ node_validator.go
@@ -1,4 +1,4 @@
-// Copyright 2013-2020 Aerospike, Inc.
+// Copyright 2014-2021 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,7 +49,9 @@ type nodeValidator struct {
 	sessionToken      []byte
 	SessionExpiration time.Time
 
-	supportsFloat, supportsBatchIndex, supportsReplicas, supportsGeo, supportsPeers, supportsLUTNow, supportsTruncateNamespace, supportsClusterStable, supportsBitwiseOps bool
+	supportsFloat, supportsBatchIndex, supportsReplicas, supportsGeo,
+	supportsPeers, supportsLUTNow, supportsTruncateNamespace, supportsClusterStable,
+	supportsBitwiseOps, supportsPartitionScans bool
 }
 
 func (ndv *nodeValidator) seedNodes(cluster *Cluster, host *Host, nodesToAdd nodesToAddT) error {
@@ -317,6 +319,8 @@ func (ndv *nodeValidator) setFeatures(features string) {
 			ndv.supportsBitwiseOps = true
 		case "cluster-stable":
 			ndv.supportsClusterStable = true
+		case "pscans":
+			ndv.supportsPartitionScans = true
 		}
 	}
 }