+++ types/error.go
@@ -1,4 +1,4 @@
-// Copyright 2013-2020 Aerospike, Inc.
+// Copyright 2014-2021 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -71,14 +71,19 @@ func NewAerospikeError(code ResultCode, messages ...string) error {
 	return AerospikeError{error: err, resultCode: code}
 }
 
-var ErrServerNotAvailable = NewAerospikeError(SERVER_NOT_AVAILABLE)
-var ErrKeyNotFound = NewAerospikeError(KEY_NOT_FOUND_ERROR)
-var ErrRecordsetClosed = NewAerospikeError(RECORDSET_CLOSED)
-var ErrConnectionPoolEmpty = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Connection pool is empty. This happens when either all connection are in-use already, or no connections were available")
-var ErrTooManyConnectionsForNode = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Connection limit reached for this node. This value is controlled via ClientPolicy.LimitConnectionsToQueueSize")
-var ErrTooManyOpeningConnections = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Too many connections are trying to open at once. This value is controlled via ClientPolicy.OpeningConnectionThreshold")
-var ErrTimeout = NewAerospikeError(TIMEOUT, "command execution timed out on client: See `Policy.Timeout`")
-var ErrUDFBadResponse = NewAerospikeError(UDF_BAD_RESPONSE, "Invalid UDF return value")
-var ErrNoOperationsSpecified = NewAerospikeError(INVALID_COMMAND, "No operations were passed to QueryExecute")
-var ErrNoBinNamesAlloedInQueryExecute = NewAerospikeError(INVALID_COMMAND, "Statement.BinNames must be empty for QueryExecute")
-var ErrFilteredOut = NewAerospikeError(FILTERED_OUT)
+var (
+	ErrServerNotAvailable             = NewAerospikeError(SERVER_NOT_AVAILABLE)
+	ErrKeyNotFound                    = NewAerospikeError(KEY_NOT_FOUND_ERROR)
+	ErrRecordsetClosed                = NewAerospikeError(RECORDSET_CLOSED)
+	ErrConnectionPoolEmpty            = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Connection pool is empty. This happens when either all connection are in-use already, or no connections were available")
+	ErrTooManyConnectionsForNode      = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Connection limit reached for this node. This value is controlled via ClientPolicy.LimitConnectionsToQueueSize")
+	ErrTooManyOpeningConnections      = NewAerospikeError(NO_AVAILABLE_CONNECTIONS_TO_NODE, "Too many connections are trying to open at once. This value is controlled via ClientPolicy.OpeningConnectionThreshold")
+	ErrTimeout                        = NewAerospikeError(TIMEOUT, "command execution timed out on client: See `Policy.Timeout`")
+	ErrUDFBadResponse                 = NewAerospikeError(UDF_BAD_RESPONSE, "Invalid UDF return value")
+	ErrNoOperationsSpecified          = NewAerospikeError(INVALID_COMMAND, "No operations were passed to QueryExecute")
+	ErrNoBinNamesAlloedInQueryExecute = NewAerospikeError(INVALID_COMMAND, "Statement.BinNames must be empty for QueryExecute")
+	ErrFilteredOut                    = NewAerospikeError(FILTERED_OUT)
+	ErrPartitionScanQueryNotSupported = NewAerospikeError(PARAMETER_ERROR, "Partition Scans/Queries are not supported by all nodes in this cluster")
+	ErrScanTerminated                 = NewAerospikeError(SCAN_TERMINATED)
+	ErrQueryTerminated                = NewAerospikeError(QUERY_TERMINATED)
+)